renamed discretizers utils
